Working fix for the log to file with PrintMode=1
Initially introduced in the commit bc113ef, it was trying to pass va_list
to a variadic function directly. This caused issue #9
(https://github.com/primesearch/mfaktc/issues/9), resulting corruption
of the integer values printed to log file.
This commit provides a fix by formatting a string in buffer first and
then replacing '\r' at the end if PrintMode set to 1.